Add wrapper for helm commands
Add wrapper for install, upgrade and lint commands. The wrapper checks
for files that are b64 encoded and attempts to decrypt them. The files
are decrypted into shared memory and those contents are then passed to
the default helm command.

Decrypted content is always deleted at the end.

Fixes OS-8957.